[PGPRO-12159] Added the output of tsv lexemes positions.
If you create an index with the operator class rum_tsvector_ops,
the positions of the lexemes will be saved as additional information.
The positions are stored in compressed form in bytea.

There is a problem that is related to the fact that in the posting tree,
additional information for the senior keys is stored in a different way,
which is why it has not yet been possible to output it. For all other
cases, the output of additional information works correctly.

Tags: rum