[PGPRO-12159] Added the output of weights.
If the index is created with the appropriate class of operators,
then in addition to the positions of the lexemes, weights (A, B, C, D)
are also stored in the additional information. Their output has been added.

In addition, Asserts have been added to the find_add_info_atr_num() and
find_add_info_oid() functions, which check that there is only one (or zero)
type of additional information in the index.

Tags: rum