[PGPRO-11599] Fix wrong results returned when order_by_attach=TRUE.
With order_by_attach=TRUE, the RUM index arranges the entries in
the posting lists (for key that have additional information) in
order of sorting by additional information. For the remaining keys,
the records are sorted by tid. The scanGetItemRegular() function
assumes that all posting lists have the same sorting, so it was
not suitable for searching in this case.

In order to fix this, scanGetItemRegular() has been redesigned.
Now it binds myAdvancePast to the key, which is sorted by
additional information. And the remaining keys are iterated over
at each iteration to find a suitable curItem.

This fix works very poorly on queries that use partial matching
when they scan a large amount of data. This is due to the fact
that sorting constantly has to be rewound and go through it over
and over again.

You can try to fix this by replacing the sorting (which is created
in collectMatchBitmap()) to the hash table (for a quick search for
the desired tid). Or by sorting altOrderKey by tid.

Tags: rum