[PGPRO-12159] Fixed the search for the key attribute number.
Previously, it was assumed that the RUM index could not contain
attributes with different types of additional information (and some
functions from PGPRO-12159 relied on this), but it is possible. If
you create an index for two columns of the tsvector type and attach
another column to one of them, it turns out that one tsvector has
the attach attribute as additional information, and the second one
will have positions.

In order to fix this, the functions that find the attribute number
of the key for which the posting tree was built were redesigned.
Now they scan the index and look for the key (and its attribute
number) for the scanned page.

Tags: rum