[PGPRO-11599] The situation of many altOrderKeys has been fixed.
A query can contain several conditions for the same column, and
if this column in the index is sorted by additional information,
then the query contains several keys, each of which is sorted
by additional information.

The scanGetItemRegular() function did not take this into account.
This commit corrects this situation.

Tags: rum